requirements.txt
torch
transformers
faiss-cpu
flask
